pyproject.toml
# 根目录配置 - 统一管理依赖和工作区

[project]
name = "research-gpt"
version = "1.0.0"
description = "Deep Research Agent - 基于 LangChain 和 DeepSeek 的智能研究助手"
authors = [
    {name = "Research Agent Team", email = "team@research-gpt.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.105.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.7.4",
    "email-validator>=2.2.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "requests>=2.31.0",
    "tavily-python>=0.3.0",
    "duckduckgo-search>=3.9.0",
    "wikipedia>=1.4.0",
    "ty>=0.0.1a24",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "asyncpg>=0.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# [tool.uv.workspace]
# members = [
#     "backend",
# ]




# [tool.uv] 工具配置
# [tool.uv]
# exclude = [
#     "tests/",
#     "scripts/",
#     "migrations/",
#     "__pycache__/",
#     ".venv/",
#     "venv/",
# ]
